Align the Windows test suite with the current API surface
The library now compiles fully on Windows; the test target was written
against a pre-rename surface and never compiled. Alignment:

- Test-module 'typealias Kernel = Windows.32.Kernel' (bare Kernel is
  the L3 unifier's name per [PLAT-ARCH-008k]; in tests it is private
  convenience) — clears every 'cannot find Kernel' site.
- Syscall-literal era names -> ISO mirrors: Unlink->File.Delete,
  Mkdir->Directory.Create, Rmdir->Directory.Remove, Rename->File.Move,
  Sync->File.Flush (sync/datasync -> flush/flushData),
  Symlink->Link.Symbolic, Dup->Descriptor.Duplicate (dup->duplicate),
  Seek->File.Seek, Link.link->Link.create(source:linkPath:).
- Error mappers gain a current(from: UInt32) testing seam (current()
  delegates) — the shape the error-mapping tests exercise.
- Mode tests -> two-Bool API; Directory keypath tests -> type-existence
  (key paths cannot root at a ~Copyable type); Entry test names
  null-terminated per the ISO convention; System/Granularity
  .rawValue -> .underlying; Memory.Map.Flags -> .Options, Protection
  combined via | (no insert), @testable import of the Memory Map module
  for the internal windowsVirtualProtect (+ test-target dep);
  Socket.Protocol reached via in-scope typealias (.Protocol is metatype
  syntax even backticked), IPPROTO enum .rawValue comparisons;
  GENERIC_READ|GENERIC_WRITE mixed-sign OR split; Thread test's
  captured-var mutation -> @unchecked Sendable flag box; Mutex
  TestError: Swift.Error (Error_Primitives shadows Error); dup2 tests
  removed (no Win32 surface).

+++ Tests/Windows 32 Kernel Tests/Windows.Kernel.IO.Read Tests.swift
@@ -81,7 +81,7 @@ extension Windows.`32`.Kernel.IO.Read.Test.Unit {
         let handle = utf16Path.withUnsafeMutableBufferPointer { pathPtr in
             CreateFileW(
                 pathPtr.baseAddress,
-                DWORD(GENERIC_READ | GENERIC_WRITE),
+                DWORD(GENERIC_READ) | DWORD(GENERIC_WRITE),
                 0,
                 nil,
                 DWORD(CREATE_ALWAYS),
